Support multiple network interfaces per container
IP addresses are not assigned to containers, but to network interfaces
that belong to containers. It's now possible to add more interfaces of
different types. Only routed veth is supported for now.

The way the container's network is configured inside has changed. osctld
no longer supports arbitrary script/programs, but the per-distribution
configuration logic is now included. This was necessary, as it would be
awkward to pass more information via environment variables or arguments
to configuration scripts.