Support for persistent container consoles
Container's console can be opened whether it is running or not. When the
container is started, it's done so using lxc-start --foreground, which
allows osctld and its users to see the container start process, i.e.
what services are starting and possibly error messages.

lxc-start -F is run in a wrapper script that runs it within a new pty.
stdin/stdout of this process is redirected to two named pipes, one for
writing, one for reading. osctld is accessing the console using these
pipes. The pipes are named so that the container can survive osctld
crash/restart.

osctl start/stop/restart has a new switch -F, --foreground, which opens
the console before executing the action. The console can be detached in
the same way as osctl console.