Store CT rootfs at lxc/ct/<id>
This should make it simpler to find the container data, as you no longer
need to know the user name. LXC configs are still located in the user's
LXC path, which has been moved to lxc/user/<name>. The container's user
is now stored in the config file, which has been moved to
`lxc/conf/ct/<id>.yml`.