Put each container in its own cgroup
These container cgroups are root-owned, so it is possible to fine-tune
cgroup parameters from outside LXC, i.e. the user himself cannot change
the cgroup values, while osctld can.

osctl ct su now scopes the shell to the container's cgroup, so you
cannot use the shell to start any other container, as that would put
it to a wrong cgroup.