osctl: refactor osctld client, response wrapper and error handling
OsCtl::Client now wraps the response from osctld within a class. One can
choose to get the response or the data it contains. There are also bang
methods that raise exceptions when errors occur. This greatly simplifies
error handling, as you don't have to check for errors manually and
simply assume that everything is ok.