Groups: add parameter full_path
`path` doesn't include the root group, which is necessary if you're to
read cgroup parameters. At the same time, the root group is somewhat
hidden, it is prefixed to group paths automatically. You don't specify
the root group when creating groups, so it'd be from to show the path
including the root group. Thus, `full_path` includes the root group.