CT veth up/down hooks: identify interfaces by their names, not indexes
Indexes change when interfaces are added/removed. For example, if you
had bridged eth0, routed eth1. eth0 had index 0, eth1 had index 1. This
index was used in hook link names. When eth0 was deleted, the hook for
eth1 was left pointing to interface with index 1, but in osctld, its
index was already 0. osctld would either have to reindex interfaces on
change and rewrite configs & hooks, or identify interfaces by their
names. This has the added benefit of being more user friendly anyway.

Indexes are still used to fetch host veth name on osctld start, but that
is alright, since when osctld is starting, interface indexes are
correct.