osctl ct mount dataset: remove --opts, use --ro/rw, mountpoint as arg
As shown in issue #4, mounting of datasets was unnecessarily complicated.
Since it is always a bind mount, it is useless to ask the user to provide
mount options other then ro/rw. These are now set using --ro and --rw
options. Option create=dir is added implicitly. The --mountpoint option
was transformed into an argument.