Relax access permissions on container's rootfs dataset/directory
In an attempt to restrict access to a container's rootfs directory only
to appropriate user namespace, it was made too strict. Only root from
the user namespace could access this directory. This meant that no other
user from the same user namespace could access it, making it impossible
to access the filesystem for anyone but root.