Use lxc.autodev=1 and bind-mount devices from the host
Apparently, systemd expects /dev to be mounted as as tmpfs or devtmpfs,
and if it isn't, it tries to mount it... which fails and the container
cannot be started. So, instead of a static /dev, osctld is populating
/dev every time the container starts, using lxc.hook.autodev.

The autodev hook is called from the container's user namespace, so a
special user-control socket had to be created,
/run/osctl/user-control/namespaced.sock. It identifies osctld user by
the client's namespaced uid/gid.