osctl user new: check for duplicate ugid offset
Since osctld is using lxc.hook.autodev, the user's offset has to be
unique. This is because the hook is run as the namespaced root user and
in order to identify him, the user is searched by the offset. If there
were more users with the same offset, we couldn't safely identify them.