osctl user: allow duplicit uid/gid mappings
Duplicit uid/gid mappings were forbidden, because osctld needed to
identify users from pre-mount and autodev hooks, which are called from
within a user namespace. It located the user using the mapping, which in
result had to be unique. We're now using the fact that osctld knows the
pid of the caller (hook), so it can find its grandparent and use its
uid within host namespace for identification.