osctl: configurable position of the container's interconnecting address
To change the position of the container's interconnecting address on
routed veth interface, simply add the address using
`osctl ct netif ip add`. osctld will then add the addresses in the
configured order. If the container's interconnecting is not visible in
`osctl ct netif ip ls`, it is added as the first address on the
interface.