osctld: use unix sockets intead of named pipes for container consoles
Opening of a named pipe blocks the caller until both sides of the pipe
are opened. When multiple containers were started simultaneously, there
was some race condition that caused osctld to deadlock due to these
blocking pipes. Using UNIX sockets instead of the pipes workarounds the
issue. In addition, containers will not not be blocked when writing to
the console if osctld is not running. Container wrappers buffer the
output and send it to osctld when it reconnects.