os: switch: stop/start changed services instead of a restart
Changed services are stopped before the new system is activated. It is
better to have the services stopped when the new system activates,
because we can be sure that when they start, their dependencies are also
up-to-date and not awaiting restart.

For example, when deploying declarative containers, some container
services were restarted before osctld was. This would be problematic
when new osctld version is really needed, but the old version is still
running.