osctl ct exec: add option --network
When --network is used, the user command will be executed after the
container's network has been setup. Only static addresses/routes are
configured, DHCP is not run. Only IPv4 has the default gateway set,
because IPv6 gateway changes on every start and we don't have the
information when generating the configuration.