os: overlays: add to nix path, load in nix-shell
`os/overlays/default.nix` contains a list of all overlays and is used
when building the OS. If `os/configs/devel.nix` is imported, the
overlays are added to NIX_PATH.  However, the default overlay list cannot
be added to the path, because it's a function that takes path to vpsadmin
as an argument. For this reason, overlays that can be used in NIX_PATH
were moved to `os/overlays/common.nix`.