osctld: fail when holding inclusive lock and acquiring exclusive lock
When a thread is holding inclusive lock and tries to acquire exclusive
lock on the same object, the thread will hang, because in order to
acquire an exclusive lock, no other lock can exist. At the same time,
inclusive locks cannot be safely promoted to exclusive locks, especially
when using `inclusively()` blocks.