os: runit: force all services to use the current system path
runsvdir and all runsv processes are started with PATH set to the path
of the current system. However, when new system versions are activated,
all services started by runsv are still using the original system path,
because runsv processes are never restarted and so the original PATH
remains in their environment and is passed to their children, the
services.