osctl-repo: download template first, stream afterwards
`osctl-repo remote get` was designed so that as it was downloading the
template, it was simultaneously being streamed to the caller via stdout.
This wasn't a good idea since when the caller is osctld, it is
extracting the downloaded template to the pool, which might be slowing
down the download.

In addition, only one process could read one template at a time. Other
processes waited on a lock. Now, the template is downloaded with a lock,
but usually the template is up-to-date. The template is then opened,
unlocked and streamed to the caller. Multiple processes can now read one
template simultaneously.