image-scripts: Alpine: symlink /sys/fs/cgroup/openrc to systemd
The systemd cgroup has to be mounted, because the majority of VPS use systemd,
which does not work without it. cgroup for openrc is rather complicated
to mount -- all VPS would have to have it mounted, otherwise some
software such as Docker would not work. So we instead reuse the systemd
cgroup for openrc purposes.