osctl ct reinstall: replace original rootfs dataset
The container rootfs dataset is now destroyed using zfs destroy, which
is faster than removing all files. In order to achieve this, a new
dataset with identical properties is created, all subdatasets are moved
to the new location and then the original dataset can be destroyed. The
new dataset is then moved in place of the original dataset.