osctl shutdown: handle osctld crash/restart during shutdown
osctld is in shutdown mode if /run/osctl/shutdown exists. When osctld
correcly prepared the system for shutdown, the executable permission
by user is set. osctl shutdown does not exit when the connection to
osctld is unexpectedly closed, but instead waits for /run/osctl/shutdown
to gain the executable by user permission.