os: disable unified cgroup hierarchy in nixos container configuration
nixpkgs unstable has switched to the cgroupv2 hieararchy by default, but
that breaks cgroups in NixOS containers on vpsAdminOS, because the host
uses cgroupv1.

For Fedora, we've dealt with this by changing init cmd in the container
image to add systemd.unified_cgroup_hierarchy=0 argument to systemd.
This does not work for NixOS, because its init does not pass given
arguments to systemd.

NixOS has option systemd.enableUnifiedCgroupHierarchy to revert to
cgroupv1, but that doesn't work in containers, because it changes kernel
parameters in /proc/cmdline -- it has no effect in a container. For this
reason, we use option boot.systemdExecutable to add the
systemd.unified_cgroup_hierarchy=0 argument to systemd.