osctld: exec while attaching into cts
When entering cts e.g. for osctl ct exec, we used to fork from osctld,
attach to the appropriate cgroups and switch to the container's user.
Then we can access LXC via API or CLI. This works, but is troublesome,
because osctld tends to have a large memory footprint and the forked
processes is entering the container's memory cgroup. This can cause
memory within the ct to run out.

So this commit adds exec() after fork and changes behaviour of
ContainerControl::Frontend#pipe_runner to pass arguments to the execed
process.