image-scripts: Add Fedora 34
NetworkManager does not manage veth interfaces by default. It needs to
be told explicitly by udev to manage such interfaces, hence the
/etc/udev/rules.d/86-vpsadminos.rules file.

The problem is that this rule is not invoked, because we mask
systemd-udev-trigger.service. This service fails due to restrictions on
/sys enforced by vpsAdminOS. Fedora 34 at least changed the service so
that it uses ExecStart=-udevadm, i.e. the service won't fail, but
udevadm will exit on restricted /sys before sending events about veth
interfaces.

We therefore override systemd-udev-trigger.service to run just
udevadm trigger --subsystem-match=net --action=add. This will invoke the
udev rule and enable the interface within NetworkManager.

This change may also be needed in the future in CentOS Stream and others
where NM is used to setup network. For now it is applied to all Fedoras.