osctld: container datasets are mounted only within containers
Container datasets are no longer mounted globally on the host.
The datasets are mounted during container start from the
lxc.hook.pre-mount hook, which is called in the container's mount
namespace. Leaving the datasets unmounted on the host required changing
osctld to fork into a new mount namespace whenever access to the
container's filesystems is needed. It is never mounted on the host,
osctld instead creates a temporary mount namespace in a forked process.
This is used mainly for distribution support.

Container datasets should now be always unmounted.