osctld: switch user before exec() in exec_runner
If we switch the user after exec to the runner process, we'll end up
with root-owned file descriptors. In order for them to be owned by the
unprivileged user, it has to be set before exec().

We need the file descriptors to belong to the user if LXC attach using
pidfd passing is to work.