os: update declarative nixos containers to work with unmounted datasets
Instead of populating the container's Nix store from the runit service
on the host, we generate configure a post-mount hook. Within this hook
the container's dataset is mounted and missing paths are added to its
store.

Since the runit service does not have access to the container's rootfs,
every container update requires the container to be restarted for the
post-mount hook to be run. This approach appears less convenient, but it
is cleaner in the sense that we do not touch the Nix store behind Nix
daemon's back.