osctld: set memory cgroup parameters twice on failure
This is to help changes of memory limits, when both
memory.limit_in_bytes and memory.memsw.limit_in_bytes is set at the same
time. The issue is, that memory.memsw.limit_in_bytes cannot be lower than
memory.limit_in_bytes. This poses a problem when the limits are changed...
the order of setting these limits depends whether we're increasing
or decreasing the limit.

As it is, the code does not easily allow for this ordering to be
implemented. This solution is less nice, but will work. At least one
of the limits will be set on the first pass and if the other fails, it
will be set on the second pass.