osctld: automatically update systemd command based on cgroup version
When cgroupv1 is used, recent systemd versions need option
systemd.unified_cgroup_hierarchy=0 as an argument to /sbin/init in order
to use the hybrid hierarchy. Older versions of systemd don't need this
option, but it should be harmless to set it there as well.

On cgroupv2 systems, we must remove the option so that the unified
hierarchy is used.

Container configs are thus adapted every time they are loaded, e.g. during
pool import, container creation, import or send/recv. This makes it
possible to send/recv containers with systemd between vpsAdminOS hosts
with different cgroup versions.