os: switch-to-configuration: activate lxcfs using osctld start config
Calling osctl activate --lxcfs after osctld is restarted is unreliable,
because it takes time to import pools and containers. The script only
waited for osctld to start, which meant that there weren't any
containers at that time or not all of them.

We now generate one-time start config for osctld before it is started.
osctld will read it and activate lxcfs if needed after all pools
and containers have been imported.