osctl ct top: redesign json exporter refreshing using a queue
Raising an exception from the signal handler worked only when the block
would run while the sleep() was running. If it was raised when the stats
were being measured and written out, the program would crash.

The only downside of this solution is that we need to run a new thread
frmo the singal handler to access the queue.

(cherry picked from commit e445ced64e30dbc6cb3a2654dcc0c3f853ac9116)