osctld: announce container init pid in a standalone event
Commit 7239cc1a88dba631c710e304c5a4e8380c97dcf6 added container init PID
to the state change event if the state was RUNNING. It did this by
sending the event only after the init PID was read from LXC using
ContainerControl. This may or may not have caused the RUNNING event to
not be announced at all, which in turn causes the start command to time
out and report error. This could happen if the init PID readout did not
finish, it is unclear whether this is the case.

(cherry picked from commit 77dd76d752f73f37d6e023ab480274742c377b65)