osctld: disable ksoftlimd on container stop
This is to solve a possible issue when the container's memory is almost
full, but ksoftlimd manages to free just enough memory for the OOM
killer to not take action. What we observe is containers being unable to
shutdown because of this.

(cherry picked from commit ac0ba4a8269c51d43df71c10fccc9ebafb2735b5)