osctld: disable automount on pre-start mounts
If automount is true, the mount manager tries to propagate it to the
container if it is running. That means he probes the container's state
while we know that it is starting and not yet running.

It is enough to have in_config: true for the mounts to appear in LXC
configuration.