osctld: wait for pool import when rebooting containers
If a container is rebooted while osctld is being restarted, we can
either miss the event entirely, or we can catch it while the pool
is being imported. When we do catch it, we must wait until the pool
is imported, because the start command will return an error.

(cherry picked from commit ae4b266fc272de69f8e2f35e7efbabc0d2e14b1c)