osctld: cpu-scheduler: do not unschedule immediately during upkeep
It is possible that the upkeep code would run just as the container is
starting, in which case its state would still be stopped, but we would
unschedule it just as it is starting. Unschedule only if the conditions
hold for more than three upkeep runs.

(cherry picked from commit fccfb25d1995bc48da8b68c1cb2b6e0907cac8ce)