osctld: ensure lxcfs is started on ct stop/start
When osctl ct stop and ct start are called in a quick succession, there
can be a race between lxcfs shutdown and wanted it to be up again. This
is because lxcfs is being stopped when the container's console closes,
but that the ct stop command returns sooner. So, make sure that the
service exists and that runit knows we want it up.