osctld: prefer filling up existing lxcfs workers
When assigning an lxcfs worker, we've been chosing the worker with the
least containers. This is irrelevant for node bootup, but as containers
get later restarted, it leads to more under-utilized lxcfs workers.
By choosing the most-used worker, we make it possible for the least-used
worker to eventually get shut down, thus reducing the number of lxcfs
instances.