vdevlog: keep a list of all vdevs, lookup devices from state
This commit resolves two issues. One, ZED reports whatever vdev path as
was added or stored by ZFS, which can be a short /dev name, but it can
also be a /dev/by-id or /dev/by-path name. By storing a list of all
vdevs and their guids, we can quickly lookup appropriate vdev using
just guid.

Second, a frequent kind of hdd errors is that the drive disconnects from
the system and then reconnects. When ZED reports IO errors, the device
might not be visible in the system, so at that time vdevlog failed to
look it up. By always having a list of all vdevs stored in the state, we
don't need the device to be present in the system when handling ZED
event.

(cherry picked from commit 978e8c9f5d51cb049e29aec34c2b9100f556fc88)