osctl-image: sync filesystem after image build
We've had issues with Alpine builds containing zeroed files. This is
because apk installs files as:

  open()
  fallocate()
  mmap()
  ...write data...
  munmap()

After that, we make a ZFS snapshot. As there is no msync(), the snapshot
contained empty files, because the page cache hasn't been written to disk
yet. Syncing the filesystem will write the caches to disk.