osctl ct top: handle negative real-time cpu usage
To gather cpu usage stats, we read /proc/stats and then counters in cpu
cgroups of all containers. These counters use different units and are
read one by one, so it is not all that precise. The more cgroups
and higher load there is, the less precise the readings are. I'm not
aware of a better way, so we hide the imprecision by showing zero
usage instead of negative results.