image-scripts: guix: Fix bootloader installation.
While installing the bootloader is not required (this is running inside a
container after all), not being able to do so implies the need to always pass
--no-bootloader flag while reconfiguring.  That complicates automation and needs
to be remembered.

The solution is to change the bootloader used to one that can be installed
inside the container.  That does waste some space (~8MB) in /boot, but users can
still opt to run the command with --no-bootloader if they would consider it to
be a problem.

* image-scripts/images/guix/vpsadminos.scm (%ct-bootloader): Use
grub-efi-netboot-removable-bootloader and install into /boot.