image-scripts: guix: Fix vpsadminos-networking when reconfiguring.
Until now the vpsadminos-networking service failed on each reconfigure, since it
could not be restarted (the addresses and routes were already configured).
Original plan was to wait until the /run is on tmpfs, but that could be a while
and the failure is annoying.

The solution is to mount a separate tmpfs as /run/vpsadminos and store the flag
in there.  In this way it will get cleaned on reboot, while not causing any
troubles for rest of the system by making whole /run a tmpfs.

* image-scripts/images/guix/vpsadminos.scm (%ct-file-systems): Mount tmpfs at
/run/vpsadminos.
(vpsadminos-networking): Use a flag file to ensure the script runs just once per
boot.