osctld: run guix halt in a thread to avoid hangs
Shepherd is not able to shut down reliably and it can happen that the
halt command run inside the container will not exit, either. In that
case, osctld was stuck on it and hard kill was never triggered. The halt
command is now run in a thread, so that it does not block the timeout
that leads to hard kill.