image-scripts: guix: Switch to linux-libre.
Due to upstream changes, the %ct-dummy-kernel approach no longer works.  Details
are in this[0] bug report.

It seems only option we have right now is to just use the libre-linux kernel.
It wastes both the disk space and CPU during the system profile build, but there
does not really seem to be any other way, at least for now.

0: https://debbugs.gnu.org/cgi/bugreport.cgi?bug=70239