os: separate nixos and vpsadminos channels in separate store paths
Individual channels must be in their own store paths and symlinked into
the resulting channel path, which is what the profile then points to.
When they're in the same store path, we can't update individual channels
using nix-channel --update, because it results in file conflicts. The
way nix-channel works, it copies unchanged channels and updates the
channels we want to update. Now, if nixos and vpsadminos channels are in
the same path, nix-channel will copy it and then it tries to add the
updated channel, which fails, because the original channel was already
copied.