osctl, osctld: add garbage collector to track temporary datasets
The garbage collector is used to track the lifetime of temporary
container datasets used for `osctl ct boot` and containers with NixOS
impermanence. These datasets will exist as long as the container needs
it and be moved to trash afterwards.

While we used to move boot datasets to trash when the container was
stopped/restarted, the solution was imperfect, as osctld was only aware
of those datasets as long as the host machine wasn't reset. The garbage
collector now tracks the datasets persistently and will find them after
system reset.

We use a new RunId facitility to identify different container runs. The
datasets' lifetime is tied to a specific run. RunId is a set of pool
name, container id and timestamp of when it was started.