osctld: stop waiting for ct to start when state is recovered
It can happen that the container fails to start in a way that the
starting process does not notice it, i.e. when the expected state
transition does not take place and the container remains stopped. osctld
can end up waiting for it to start while holding its lock for eternity
and there was no way to recover except for osctld restart. This happens
almost exclusively when something is wrong on the machine, e.g.
processes stuck in the kernel, etc.

Now, when the administrator will run `osctl ct state recover` and the
container is found to be stopped, the waiting process will abort.