osctld: fix dataset trashing when prune is running
It could happen that prune is triggered when dataset is being added to
trash. Adding the dataset to trash consists of zfs rename and zfs set.
Prune was free to destroy the datasets between those two commands,
leading to zfs set failing.

Prune now checks that the properties on the trashed dataset are set
before destroying it.