osctld: do not force-start during reboot
When reboot is called, the container's state already has to be set to
`stopped`. Forcing it can potentionally lead to initializing the
container twice if the user uses both reboot from the inside and start
from the outside. The reboot code waited for the manipulation lock and
forcefully attempted to start it.