osctld: announce state running after init_pid is set
This solves issue with test runs in osctl-image. When the test container
was being torn down, it was first killed and then deleted. However, the
test that only starts the container was done before init_pid was set on
that container. Since osctl-image uses Ruby client to send commands to
osctld, it was faster than osctld's internal procedure to read the
container's init_pid.

When container kill was invoked before init_pid was set, the stop
command did not wait for the exit promise. That led to a race condition
between exit handling in console/console.rb and container deletion,
which is run immediately after it is killed. The exit handler tried to
remount the dataset to write-out dirtied pages while the delete command
was already running and moving the dataset to trash.

Announcing `running` only after init_pid is set will make the stop
command wait for the exit promise and thus avoid the race.