osvm: fix timeout handling in Machine#wait_for_osctl_pool
wait_for_osctl_pool left the timeout checking to wait_until_succeeds,
which it runs in a loop. However, that works only when the executed
command fails. When the osctl pool exists but isn't active yet, the
command was always succesful and timeout was never hit, e.g. when the
pool was stuck in 'importing' state.