osvm, tests: use per-machine disks
When a test included multiple machines that were based on tank.nix or
with-tank.nix, they all used one disk file. That doesn't work when the
machines are run at the same time. `{machine}` in disk file path is now
replaced with machine name, which will allow machines to have their own
disks.