image-repository, tests: revert back to versions instead of old tags
stable/oldstable make sense only for Debian, where those codenames are
actually used. For other distributions it is rather confusing and since
the support times are long, even oldold might not be enough. Therefore,
revert back to using versions for Alma, Rocky and Ubuntu, even though it
will require changes to distributions.nix to keep it up to date.