test-runner, tests: support for multiple scripts per test
Test templates generate tests that are run in parallel independent of
each other. Each test can now have multiple test scripts, which are run
one by one. The order in which they are run is not guaranteed. All test
scripts of one test share the execution environment and virtual machines.

A test is successful if all its scripts have the expected result.