os: separate secrets from Nix store in rootfs squashfs image
Currently, the rootfs squash image contains directly the contents of
/nix/store, not /nix/store itself, because mksquashfs implicitly adds
base names of listed directories/files. This means that within the
image, the secrets directory is on the same level as store paths.

Since we do not want the secrets to be world-readable from /nix/store,
stage-2 removed /nix/store/secrets from the read-write overlay that is
mounted to /nix/store. Now that we preserve the original read-only layer
for mounting into declarative containers with shared store, the secrets
would be accessible within the containers' /nix/store overlay and in
/nix/.ro-store on the host.

By adding -no-strip to mksquashfs, the squash image will contain:

  - /nix/store
  - /secrets

/secrets is copied to /var/secrets in stage-1 and cannot be accessed
from /nix/store anymore.