Add osctl-oomd
osctl-oomd is used to restart/stop containers that are overloading the
host system by running out of memory. Containers are targetted by:

  - CPU usage close to the configured limit
  - memory usage close to the configured limit
  - it must be among the containers with the highest load average

In addition, osctl-oomd is active only when the host's 1 minute load
average exceeds a multiple of load average median.

When all these conditions are met for a configured period of time, the
abusing containers are restarted and later stopped if restart won't
help.