github: include kernel.dev within CI builds
vpsfree-cz-configuration uses `config.boot.kernelPackages.perf`, which
requires `kernel.dev` to be built. As perf is not included in vpsAdminOS
by default and `kernel.dev` isn't in the binary cache, kernel rebuild
was triggered. By adding `kernel.dev` to the top-level closure, we
ensure that it is always included in the binary cache.