osctld: use fe80::1 as default IPv6 gateway for routed veth
osctld used the auto-generated link-local address from the host-side
interface. However, that address is different on each container start,
because MAC addresses on the host side are auto-generated and cannot be
configured statically in LXC.

When the veth interfaces comes up, osctld adds fe80::1/64 address to it,
i.e. each veth interface on the host has this address. Containers then
use this address as a gateway. This allows the container users to configure
network manually.