os: use zfs share -a instead of -r
zfs share -r is provided by our patch to OpenZFS, which doesn't work as
expected. The idea was to *add* exports of the specified pool and no
other pools. In reality, ZFS seemingly flushes all exports and then adds
exports of the specified pool, thus rendering exports from other pools
unavailable. This will be difficult to fix in ZFS and we might as well
drop the patch and use zfs share -a.