os: enable systemd-sysctl.service in NixOS containers
In the past, the service used to fail, because it tried to set sysctls
that are not available within a container. However, there are sysctls
that can be set from within a container that users might want to use
and since at least in the default configuration the service no longer
fails, let us enable it.